+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+# v0.9.6
+## 16-06-2017
+
+1. [](#improved)
+    * Optimize CSS
+    * Merge config with page, clean up Utilities
+    * Clean up Utilities
+
 # v0.9.5
-## xx-06-2017
+## 14-06-2017
 
 1. [](#new)
     * Add settings for heading font and block-text font